demo-app/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
ultralytics==8.3.170
torch>=2.5.0
pillow==11.3.0
numpy==2.3.2
pydantic==2.11.7
pytest==8.4.1
httpx==0.28.1
